+++ shell/key-bindings.zsh
@@ -103,8 +103,8 @@ fzf-history-widget() {
   local ret=$?
   if [ -n "$selected" ]; then
     num=$(awk '{print $1}' <<< "$selected")
-    if [[ "$num" =~ ^[1-9][0-9]*$ ]]; then
-      zle vi-fetch-history -n $num
+    if [[ "$num" =~ '^[1-9][0-9]*\*?$' ]]; then
+      zle vi-fetch-history -n ${num%\*}
     else # selected is a custom query, not from history
       LBUFFER="$selected"
     fi